requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.1
flask_cors==6.0.1
langchain==0.3.26
langchain_community==0.3.27
langchain_google_genai==2.1.6
google-ai-generativelanguage==0.6.18
protobuf==6.31.1
python-dotenv==1.1.1
requests==2.32.4
yfinance==0.2.65
faiss-cpu==1.8.0
tqdm==4.66.4
pandas==2.2.2
numpy==1.26.4
gunicorn==23.0.0
google-genai
duckduckgo_search==8.1.0
selenium==4.29.0